updated errors, changed authentication settings for product CRUD

+++ routes/web.php
@@ -40,12 +40,12 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::get('/product', [ProductController::class, 'index'])->name('product.index');
-Route::get('/product/create', [ProductController::class, 'create'])->name('product.create');
-Route::post('/product', [ProductController::class, 'store'])->name('product.store');
-Route::get('/product/{product}/edit', [ProductController::class, 'edit'])->name('product.edit');
-Route::put('/product/{product}/update', [ProductController::class, 'update'])->name('product.update');
-Route::delete('/product/{product}/destroy', [ProductController::class, 'destroy'])->name('product.destroy');
-Route::get('/product/{product}/show', [ProductController::class, 'show'])->name('product.show');
+Route::get('/product', [ProductController::class, 'index'])->name('product.index')->middleware(['auth', 'admin']);
+Route::get('/product/create', [ProductController::class, 'create'])->name('product.create')->middleware(['auth', 'admin']);
+Route::post('/product', [ProductController::class, 'store'])->name('product.store')->middleware(['auth', 'admin']);
+Route::get('/product/{product}/edit', [ProductController::class, 'edit'])->name('product.edit')->middleware(['auth', 'admin']);
+Route::put('/product/{product}/update', [ProductController::class, 'update'])->name('product.update')->middleware(['auth', 'admin']);
+Route::delete('/product/{product}/destroy', [ProductController::class, 'destroy'])->name('product.destroy')->middleware(['auth', 'admin']);
+Route::get('/product/{product}/show', [ProductController::class, 'show'])->name('product.show')->middleware(['auth', 'admin']);
 
 require __DIR__.'/auth.php';
